pyproject.toml
[tool.poetry]
name = "finance-bot"
version = "0.1.0"
description = ""
authors = ["Ilya Osipov <070809010@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "2.0.1"
numpy = "1.24.3"
yfinance = "0.2.18"
python-telegram-bot = "^13.1"
pytest = "7.3.1"
pytest-mock = "3.10.0"
pydantic = {extras = ["dotenv"], version = "1.10.8"}
flake8 = "6.0.*"
flake8-cognitive-complexity = "0.1.*"
flake8-commas = "2.1.*"
flake8-builtins = "2.1.*"
flake8-docstrings = "1.7.*"
mypy = "1.3.0"
pandas-stubs = "2.0.1.230501"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
